+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 ### Added
+- Adds inbound and outbound TLS support for gRPC. See `gprc.InboundCredentials`,
+  `grpc.DialerCredentials`, and `grpc.Transport.NewDialer` for usage.
 - Added `peer/peerlist/v2` which differs from the original `peer/peerlist` by
   replacing the use of `api/peer.ListImplementation` with
   `peer/peerlist/v2.Implementation`, which threads the peer separately from the